+++ lib/features/movie/presenter/movie_recomended/presenter/bloc/get_favorite_list.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'package:kueski_challenge/core/utils/status.dart';
 import 'package:kueski_challenge/features/movie/domain/entity/movie_entity.dart';
 import 'package:kueski_challenge/features/movie/domain/repository/movie_repository.dart';
@@ -10,17 +11,17 @@ class GetFavoriteList extends ChangeNotifier {
     required MovieRepository movieRepository,
   }) : _movieRepository = movieRepository;
 
-  static int currentPage = 0;
+  static int currentPage = 1;
   static int totalPages = 0;
   List<MovieEntity> items = [];
   List<int> cpMovies = [];
   ({Status status, List<MovieEntity> movies}) status =
       (status: Status.init, movies: <MovieEntity>[]);
 
-  Future<void> getList() async {
+  Future<void> getList({required bool isListener}) async {
     status = (status: Status.init, movies: <MovieEntity>[]);
-    final either =
-        await _movieRepository.getFavoritesMovies(page: currentPage += 1);
+    final either = await _movieRepository.getFavoritesMovies(
+        page: isListener ? currentPage += 1 : currentPage);
 
     either.when(
       (statusCode, error) =>
@@ -28,9 +29,11 @@ class GetFavoriteList extends ChangeNotifier {
       () => status = status = (status: Status.loading, movies: <MovieEntity>[]),
       (response) {
         totalPages = response.totalPages;
-        items = response.results;
+
+        items.addAll(response.results);
+
         cpMovies = items.map((e) => e.id).toList();
-        status = (status: Status.success, movies: items);
+        status = (status: Status.success, movies: items.toSet().toList());
       },
     );