+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 V: V1.0.0
 NOTES: Release app with target features
 Breaking Change: None
+
+V: V2.0.0
+
+- feat: Support Favorites
+- feat: Add screen advice
+- feat: favorites movies